Kucoin-volatility-trading-bot-main/requirements_websocket.txt
kucoin-python
PyYAML
requests==2.25.1
SQLAlchemy==1.4.25
sqlalchemy-utils
psycopg2-binary
pandas
